+++ Packages/BrandonUtils/CHANGELOG.md
@@ -7,6 +7,20 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 > **Note:** Dates are in `mm/dd/yyyy` format.
 
+## [2.35.1]
+
+### Added
+
+- `Mathb`
+  - `IsWhole()`
+  - `ShiftDecimal()`
+  - `Squared()`
+  - `Cubed()`
+  - `Pow()`
+  - `AsInt()`
+  - `AsLong()`
+- `LambdaExtensions.Invoke(Tuple)`
+
 ## [2.34.1] - 10/20/21
 
 ### Added